refactor: add insert_locale_mapping and move Locale to its own file
