+++ topgg/locale.py
@@ -1,0 +1,24 @@
+from enum import Enum
+
+
+class Locale(Enum):
+  """A supported locale."""
+
+  __slots__: tuple[str, ...] = ()
+
+  ENGLISH = 'en'
+  GERMAN = 'de'
+  FRENCH = 'fr'
+  PORTUGUESE = 'pt'
+  TURKISH = 'tr'
+  HINDI = 'hi'
+  JAPANESE = 'ja'
+  ARABIC = 'ar'
+  DUTCH = 'nl'
+  KOREAN = 'ko'
+  ITALIAN = 'it'
+  SPANISH = 'es'
+  RUSSIAN = 'ru'
+  UKRAINIAN = 'uk'
+  VIETNAMESE = 'vi'
+  CHINESE = 'zh'
